Improved container deployment and configuration with a few file updates
Changes:
- `Dockerfile`: explicitly copy .env.production
file to image.
- `deploy_to_new_instance.sh`: improved deployment
automation - added checks before installing some
packages, and added general comments to the file.
- `docker-compose.yml`: updated configurations -
removed volumes in the app container.

+++ deploy_to_new_instance.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/bash
 
+# Before running the script, there are a number of conditions necessary for its operation: 1. This script and the .env file with the necessary variable values ​​for the correct launch and operation of the application must be in the root folder of the user with root rights. 2. Before running, make sure that the .env file is present. 3. If necessary, make changes to the configuration of the relevant Nginx files to ensure that the application works correctly in the browser.
+
 # Get the absolute path to the current directory
 LOGDIR=$(pwd)
 
@@ -24,18 +26,32 @@ log "=== Starting deployment process ==="
 log "Updating package list..."
 sudo apt update && sudo apt upgrade -y | tee -a "$LOGFILE" || error_exit "Unable to update system."
 
-log "Installing Git..."
-sudo apt install -y git | tee -a "$LOGFILE" || error_exit "Failed to install Git."
+log "Checking Git installation..."
+if ! dpkg -l | grep -qw git; then
+    log "Installing Git..."
+    sudo apt install -y git | tee -a "$LOGFILE" || error_exit "Failed to install Git."
+else
+    log "Git is already installed, skipping."
+fi
 
-log "Installing Docker..."
-sudo apt install -y docker.io | tee -a "$LOGFILE" || error_exit "Failed to install Docker."
+log "Checking Docker installation..."
+if ! dpkg -l | grep -qw docker.io; then
+    log "Installing Docker..."
+    sudo apt install -y docker.io | tee -a "$LOGFILE" || error_exit "Failed to install Docker."
+    log "Starting Docker..."
+    sudo systemctl start docker | tee -a "$LOGFILE" || error_exit "Failed to start Docker."
+    sudo systemctl enable docker | tee -a "$LOGFILE" || error_exit "Failed to enable Docker on system startup."
+else
+    log "Docker is already installed, skipping."
+fi
 
-log "Starting Docker..."
-sudo systemctl start docker | tee -a "$LOGFILE" || error_exit "Failed to start Docker."
-sudo systemctl enable docker | tee -a "$LOGFILE" || error_exit "Failed to enable Docker on system startup."
-
-log "Installing Docker Compose..."
-sudo apt install -y docker-compose | tee -a "$LOGFILE" || error_exit "Failed to install Docker Compose."
+log "Checking Docker Compose installation..."
+if ! dpkg -l | grep -qw docker-compose; then
+    log "Installing Docker Compose..."
+    sudo apt install -y docker-compose | tee -a "$LOGFILE" || error_exit "Failed to install Docker Compose."
+else
+    log "Docker Compose is already installed, skipping."
+fi
 
 log "Cloning repository from GitHub..."
 if [ ! -d "NestJS-backend" ]; then
@@ -47,12 +63,12 @@ fi
 log "Changing to project directory..."
 cd NestJS-backend || error_exit "Failed to change to project directory."
 
-# Move .env file to project directory
-log "Moving .env file to the project directory..."
+# Move and rename .env file to project directory
+log "Moving .env file to the project directory and renaming..."
 if [ -f "/home/ubuntu/.env" ]; then
-    sudo mv /home/ubuntu/.env.production . | tee -a "$LOGFILE" || error_exit "Failed to move .env.production file to project directory."
+    sudo mv /home/ubuntu/.env .env.production | tee -a "$LOGFILE" || error_exit "Failed to move .env.production file to project directory."
 else
-    log "WARNING: .env.production not found in /home/ubuntu, skipping move. Check for this file in project root."
+    log "WARNING: .env not found in /home/ubuntu, skipping move. Check for .env.production file in project root."
 fi
 
 # Set permissions for .env.production
@@ -61,9 +77,9 @@ sudo chown ubuntu:ubuntu .env.production
 sudo chmod 600 .env.production | tee -a "$LOGFILE" || error_exit "Failed to set permissions for .env.production."
 
 log "Starting Docker Compose..."
-sudo --preserve-env docker-compose --env-file .env.production -f docker-compose.yml up -d --build --no-cache| tee -a "$LOGFILE" || error_exit "Failed to start Docker Compose."
+sudo --preserve-env docker-compose --env-file .env.production -f docker-compose.yml up -d --build | tee -a "$LOGFILE" || error_exit "Failed to start Docker Compose."
 
 log "Checking running containers..."
 sudo docker ps | tee -a "$LOGFILE" || error_exit "Failed to check running containers."
 
-log "=== Deployment completed successfully ==="
+log "=== Deployment completed ==="
